% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helper.R
\name{expand_effects}
\alias{expand_effects}
\title{Use configuration and effects vector to get each founder's effect, even when there are fewer alleles than founders}
\usage{
expand_effects(effects, enumerated_configuration)
}
\description{
Use configuration and effects vector to get each founder's effect, even when there are fewer alleles than founders
}
