% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_probs.R
\name{reduce_probs}
\alias{reduce_probs}
\title{Reduce elements of genoprobs list by allelic series}
\usage{
reduce_probs(probs, allelic_series)
}
\arguments{
\item{probs}{a list containing founder allele probabilities for a collection of markers}

\item{allelic_series}{a binary matrix of 0s and 1s indicating the allelic series}
}
\value{
an object with class calc_genoprob.
}
\description{
Reduce elements of genoprobs list by allelic series
}
\details{
The output has two classes: list and calc_genoprob. Class 'calc_genoprob' is needed for input to qtl2::scan1, for example. Output length is same as that of inputted allele probs object.
}
