% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_probs.R
\name{reduce_probs_array}
\alias{reduce_probs_array}
\title{Reduce a genoprobs array by allelic series}
\usage{
reduce_probs_array(array, allelic_series)
}
\arguments{
\item{array}{a probs array (not a list)}

\item{allelic_series}{a matrix of 0s and 1s to indicate allelic series}
}
\value{
a reduced array
}
\description{
Reduce a genoprobs array by allelic series
}
\details{
The function multiplies every marker's probs matrix (from array inputted) by allelic_series and returns an array.
}
