% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1.R
\name{sim1}
\alias{sim1}
\title{Simulate one trait with independent errors}
\usage{
sim1(aprobs, allelic_series, allelic_effects, error_variance = 1)
}
\arguments{
\item{aprobs}{founder allele probabilities matrix}

\item{allelic_series}{a binary matrix of 0s and 1s indicating the allelic series}

\item{allelic_effects}{a vector of effects for the functional alleles}

\item{error_variance}{error variance for the traits}
}
\value{
a vector of values for a single trait
}
\description{
Simulate one trait with independent errors
}
