% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{update_effect}
\alias{update_effect}
\title{Update one effect from the vector of effects}
\usage{
update_effect(
  effects,
  effect_index,
  distance = 1,
  trait,
  residual_variance,
  collapsed_genotypes
)
}
\arguments{
\item{effects}{the current value of the effects vector, length at least 2}

\item{distance}{parameter for proposal, ie, max distance from current effect value}

\item{trait}{univariate trait for mapping}

\item{residual_variance}{current value of residual variance.}

\item{collapsed_genotypes}{genoprobs matrix, collapsed per current configuration matrix}

\item{effects_index}{an integer from 1 to length of current effects vector, indicating which entry to update}
}
\value{
a list of length two
}
\description{
Update one effect from the vector of effects
}
