% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{update_residual_variance}
\alias{update_residual_variance}
\title{Update residual variance for Jannink & Wu 2003 MCMC}
\usage{
update_residual_variance(
  residual_variance,
  distance = 0.2,
  trait,
  collapsed_genotypes,
  effects
)
}
\arguments{
\item{residual_variance}{current value of residual variance}

\item{distance}{max distance between current value and proposed value of residual variance}

\item{trait}{a univariate trait for QTL mapping}

\item{collapsed_genotypes}{genoprobs matrix collapsed per current value of configuration matrix}

\item{effects}{current value of effects vector.}
}
\description{
Update residual variance for Jannink & Wu 2003 MCMC
}
\references{
Jannink & Wu (2003). \url{https://www.ncbi.nlm.nih.gov/pubmed/12872915}
}
