% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{jannink_mcmc}
\alias{jannink_mcmc}
\title{Draw samples from a MCMC chain per Jannink & Wu 2003}
\usage{
jannink_mcmc(initial_values, niter = 10000)
}
\arguments{
\item{initial_values}{a named list of length 3, where names are 'configuration', 'effects', and 'residual_variance'.}

\item{niter}{number of iterations}

\item{genoprobs}{a genoprobs matrix for QTL}

\item{trait}{a univariate trait affected by the QTL}

\item{residual_variance}{residual variance in the linear model}
}
\value{
a list of length two, where each entry is itself a list of results from each iteration
}
\description{
Draw samples from a MCMC chain per Jannink & Wu 2003
}
