% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{update_allelic_number}
\alias{update_allelic_number}
\title{Update allelic number (and, of course, configuration, too) in Jannink and Wu 2003 MCMC.}
\usage{
update_allelic_number(
  configuration,
  genoprobs,
  effects,
  trait,
  residual_variance
)
}
\arguments{
\item{configuration}{an allelic series matrix, ie, binary matrix of 0s and 1s mapping founders to alleles at a single QTL}

\item{genoprobs}{genotype probabilities matrix for a single marker, ie, n_subjects by n_founders at 1 marker}
}
\description{
Update allelic number (and, of course, configuration, too) in Jannink and Wu 2003 MCMC.
}
